mental-helth new_2 - Copy (2) - Copy/requirements.txt
flask
sqlalchemy
flask-sqlalchemy
python-docx
fpdf2
reportlab
sentence-transformers
scikit-learn
numpy
tqdm
python-dotenv
